+++ nextdns/provider.go
@@ -33,7 +33,6 @@ func Provider() *schema.Provider {
 			"nextdns_denylist":         resourceNextDNSDenylist(),
 			"nextdns_parental_control": resourceNextDNSParentalControl(),
 			"nextdns_privacy":          resourceNextDNSPrivacy(),
-			"nextdns_profile":          resourceNextDNSProfile(),
 			"nextdns_rewrite":          resourceNextDNSRewrite(),
 			"nextdns_security":         resourceNextDNSSecurity(),
 			"nextdns_settings":         resourceNextDNSSettings(),